# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AroundU API
  description: |
    AroundU is a local services marketplace connecting clients with skilled workers.
    
    ## Authentication
    Most endpoints require a JWT Bearer token. Obtain one via `POST /api/v1/auth/login`.
    
    ## Rate Limiting
    Certain endpoints enforce rate limits. Exceeding the limit returns `429 Too Many Requests`.
    
    ## Pagination
    List endpoints accept `page` (0-based) and `size` (max 100) query parameters.
    Responses include `totalElements`, `totalPages`, and `last` metadata.
    
    ## Sorting
    Sortable endpoints accept `sortBy`, `sortDirection`, `secondarySortBy`, and `secondarySortDirection`.
    Distance-based sorting is available via `sortByDistance=true` with latitude/longitude coordinates.
  version: 1.0.0
  contact:
    name: AroundU Dev Team
    email: support@aroundu.com
    url: https://aroundu.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:20232
    description: Local development
  - url: https://preprod.aroundu.com
    description: Pre-production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token obtained from POST /api/v1/auth/login
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Success
        data:
          type: object
      required: [success, message]
...
    AddressDTO:
      type: object
      required: [country, postalCode]
      properties:
        id:
          type: integer
          format: int64
        country:
          type: string
          enum: [US, GB, IN, CA, AU, DE, FR, JP]
          description: ISO country code
          example: US
        postalCode:
          type: string
          example: "10001"
        city:
          type: string
          example: New York
        area:
          type: string
          example: Manhattan
        latitude:
          type: number
          format: double
          example: 40.7128
        longitude:
          type: number
          format: double
          example: -74.006
        fullAddress:
          type: string
          maxLength: 500
          example: "123 Main St, Manhattan, New York, NY 10001"
...
    SkillDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        skillName:
          type: string
          example: Plumbing
    VerificationStatusDTO:
      type: object
      properties:
        isVerified:
          type: boolean
        verifiedAt:
          type: string
          format: date-time
        expiryDate:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Auth ──────────────────────────────────────────────────────
...
    WorkerSignupRequestDTO:
      type: object
      required: [name, email, phoneNumber, password, currency, currentAddress]
      properties:
        name:
          type: string
          example: Jane Worker
        email:
          type: string
          format: email
          example: jane.worker@example.com
        phoneNumber:
          type: string
          example: "+1234567891"
        password:
          type: string
          format: password
          example: WorkerP@ss456
        currency:
          type: string
          example: USD
        currentAddress:
          $ref: '#/components/schemas/AddressDTO'
        skillIds:
          type: array
          items:
            type: string
          example: ["1", "3", "5"]
    WorkerUpdateRequestDTO:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Jane Updated
        email:
          type: string
          format: email
        phoneNumber:
          type: string
          minLength: 7
          maxLength: 15
        profileImageUrl:
          type: string
        experienceYears:
          type: integer
          minimum: 0
          example: 5
        certifications:
          type: string
          maxLength: 500
          example: Licensed Plumber, EPA Certified
        isOnDuty:
          type: boolean
          example: true
        payoutAccount:
          type: string
          example: acct_123456789
        currency:
          type: string
          example: USD
    WorkerDetailDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        name:
          type: string
          example: Jane Worker
        email:
          type: string
          format: email
        phoneNumber:
          type: string
        currency:
          type: string
        currentAddress:
          $ref: '#/components/schemas/AddressDTO'
        verificationStatus:
          $ref: '#/components/schemas/VerificationStatusDTO'
        profileImageUrl:
          type: string
        skills:
          type: array
          items:
            $ref: '#/components/schemas/SkillDTO'
        engagedJobIds:
          type: array
          items:
            type: integer
            format: int64
        experienceYears:
          type: integer
        certifications:
          type: string
        isOnDuty:
          type: boolean
        payoutAccount:
          type: string

    # ── Job ───────────────────────────────────────────────────────
...
paths:
...
  /api/v1/worker/register:
    post:
      tags: [Worker]
      summary: Register worker
      description: Creates a worker account. No authentication required.
      operationId: registerWorker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerSignupRequestDTO'
      responses:
        '201':
          description: Worker registered
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        type: string
                        example: Worker registered successfully
        '400':
          description: Validation error
  /api/v1/worker/{workerId}:
    get:
      tags: [Worker]
      summary: Get worker details
      description: Fetch worker profile by ID. Accessible by the worker themselves or ADMIN.
      operationId: getWorkerDetails
      security:
        - bearerAuth: []
      parameters:
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Worker found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/WorkerDetailDTO'
        '403':
          description: Forbidden
        '404':
          description: Worker not found
    delete:
      tags: [Worker]
      summary: Delete worker
      description: Deletes worker and related data. ADMIN or the worker themselves.
      operationId: deleteWorker
      security:
        - bearerAuth: []
      parameters:
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Worker deleted
        '403':
          description: Forbidden
        '404':
          description: Worker not found
  /api/v1/worker/me:
    get:
      tags: [Worker]
      summary: Get current worker details
      description: Fetch authenticated worker's own profile.
      operationId: getMyWorkerDetails
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Worker profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/WorkerDetailDTO'
  /api/v1/worker/all:
    get:
      tags: [Worker]
      summary: List all workers (admin)
      description: Paginated listing of all workers. ADMIN only.
      operationId: getAllWorkers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
        - name: size
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Page of workers
        '403':
          description: Forbidden
  /api/v1/worker/update/{workerId}:
    patch:
      tags: [Worker]
      summary: Update worker details
      description: Partial update of worker profile fields.
      operationId: updateWorkerDetails
      security:
        - bearerAuth: []
      parameters:
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerUpdateRequestDTO'
      responses:
        '200':
          description: Worker updated
        '400':
          description: Validation error
        '403':
          description: Forbidden
        '404':
          description: Worker not found

  # ════════════════════════════════════════════════════════════════
  #  JOBS
  # ════════════════════════════════════════════════════════════════
...
  /api/v1/jobs/worker/{workerId}/{jobId}:
    get:
      tags: [Jobs]
      summary: Worker job detail
      description: Visible if job is open or assigned to the worker.
      operationId: getJobForWorker
      security:
        - bearerAuth: []
      parameters:
        - name: workerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Job detail
        '403':
          description: Forbidden
        '404':
          description: Not found

  # ════════════════════════════════════════════════════════════════
  #  JOB CODES
  # ════════════════════════════════════════════════════════════════
...